+++ src/ops/math.cpp
@@ -3,7 +3,7 @@
 
 // Integer Addition Kernel
 // The compiler sees this loop and turns it into hyperoptimised instructions with -O3 flag
-void add_I(int64_t* res, const int64_t* a, const int64_t* b, int64_t n) {
+void add_I(int64_t* res, const int64_t* a, const int64_t* b, int32_t n) {
     for (int64_t i = 0; i < n; ++i) {
         res[i] = a[i] + b[i];
     }
@@ -11,7 +11,7 @@ void add_I(int64_t* res, const int64_t* a, const int64_t* b, int64_t n) {
 
 // Float Addition Kernel
 // Also becomes hyperoptimised with -O3 flag
-void add_F(double* res, const double* a, const double* b, int64_t n) {
+void add_F(double* res, const double* a, const double* b, int32_t n) {
     for (int64_t i = 0; i < n; ++i) {
         res[i] = a[i] + b[i];
     }
